extract subnet extracting logic as a shared util
